refactor(agent): move create_agent node bodies into nodes.py
create_agent.py had grown to 686 lines, most of it ~18 nested node
closures defined inside the factory. Those bodies move verbatim into a
new nodes.py as module-level functions and make_* factories, each taking
exactly what its closure captured; create_agent.py is now graph assembly
only.

One deliberate dedup: the identical retrieval-result parsing that was
copy-pasted into select_tools and aselect_tools is now the shared
_split_retrieval_result helper.

Two node patch targets in the agent tests follow execute_hooks to its
new module. No behavior change.

Entire-Checkpoint: 845dc463e1cc
